raspberryPi/requirements.txt
bluepy==1.3.0
paho_mqtt==1.6.1
